hijo queda zombie padre dormido.

+++ TP_3/ejercicio_07.c
@@ -36,41 +36,73 @@
 #include <sys/types.h>
 #include <signal.h>
 #include <unistd.h>
+#include <wait.h>
+#include <stdlib.h>
 void manejame_esta_papa(int sig);
 void manejame_esta_nene(int sig);
 
-int i;
+int i=1;
 
 int main(int argc, char **argv)
 {
 	pid_t hijo;
+	printf("Soy dios creo un hijo\n");
 	if ((hijo = fork()) == 0){
 		//do something kid
-		
+		printf("veo si manejo la señal\n");
 		if (signal(SIGUSR1, manejame_esta_nene) == SIG_ERR){
 			perror("Error en el manejo de la señal :'(");
 			exit(1);
-		}else{
-			printf("Hijo: %d", contador);
 		}
+		printf("bucle infinito\n");
 		while (1){
 			pause();
 		}
-		//kill(getppid(), SIGUSR1);
+		exit(0);
+	}else{
+	//do something dad
+		printf("entro en padre. %d\n", i);
+		printf("mando SIGUSR1");
+		kill(hijo, SIGUSR1);
+		if (signal(SIGUSR1, manejame_esta_papa)== SIG_ERR){
+			perror("Papá no sabe manejar una señal :(");
+			exit(1);
+		}
+		
+		while (1){
+			
+			pause();
+		}
 		
 	}
-	//do something dad
-	//pause();
-	signal(SIGUSR1, manejame_esta);
-	//printf("padre\n");
-	kill(getppid(), SIGUSR1);
-	/*for(i=1;i<=20;i++){
-		printf("%d\n",i);
-	}
-	scanf("%d", &i);*/
+	wait(NULL);
+	printf("llego al final\n");
 	return 0;
-}
+}//fin main
 
 void manejame_esta_papa(sig){
-	printf("que haces\n");
+	if (signal(SIGUSR1, manejame_esta_papa)== SIG_ERR){
+		perror("Error en manejo.");
+		exit(1);
+	}
+	while (((i%2)!=0)&&(i<=20)){
+		printf("Padre: %d\n", i);
+		i++;
+	}
+	exit(0);
+}
+
+void manejame_esta_nene(sig){
+	static int contador = 2;
+	printf("Entre en el manejador del hijo. cont: %d\n", contador);
+	if ( signal(SIGUSR1, manejame_esta_nene) == SIG_ERR) {
+      perror("Error instalando manejadorHijo - ");
+      exit(-1);
+	}
+	while (((contador % 2) == 0)&&(contador <= 20)){
+		printf("Hijo: %d\n", contador);
+		contador+=2;
+		kill(getppid(), SIGUSR1);
+	}
+	exit(0);	
 }
